Create MD5 files for each FASTQ, and check

+++ ena_fetch/get_ENA_project_fastq.py
@@ -9,9 +9,16 @@ project = "PRJEB2896"
 
 # This was simple, but does not work anymore:
 # fastq_url = "http://www.ebi.ac.uk/ena/data/view/reports/sra/fastq_files/internal/%s" % project
-fastq_url = "https://www.ebi.ac.uk/ena/data/warehouse/filereport?accession=PRJEB2896&result=read_run&fields=study_accession,sample_accession,secondary_sample_accession,experiment_accession,run_accession,tax_id,scientific_name,instrument_model,library_layout,fastq_ftp,fastq_galaxy,submitted_ftp,submitted_galaxy,sra_ftp,sra_galaxy,cram_index_ftp,cram_index_galaxy&download=txt"
+# This is the default column set:
+# fastq_url = "https://www.ebi.ac.uk/ena/data/warehouse/filereport?accession=PRJEB2896&result=read_run&fields=study_accession,sample_accession,secondary_sample_accession,experiment_accession,run_accession,tax_id,scientific_name,instrument_model,library_layout,fastq_ftp,fastq_galaxy,submitted_ftp,submitted_galaxy,sra_ftp,sra_galaxy,cram_index_ftp,cram_index_galaxy&download=txt"
 # Or goto https://www.ebi.ac.uk/ena/data/view/PRJEB2896 and click on "TEXT" download link
-fastq_file = "%s_fastq.tsv" % project
+#
+# This includes important metadata including the fastq_md5 information,
+fields = "study_accession,sample_accession,secondary_sample_accession,experiment_accession,run_accession,tax_id,scientific_name,instrument_model,library_name,nominal_length,library_layout,experiment_title,study_title,study_alias,experiment_alias,run_alias,fastq_md5,fastq_ftp,submitted_md5,submitted_ftp,sra_md5,sra_ftp,cram_index_ftp".split(",")
+
+fastq_url = 'https://www.ebi.ac.uk/ena/data/warehouse/filereport?accession=%s&result=read_run&fields=%s&download=txt' % (project, ",".join(fields))
+
+fastq_file = "%s_metadata.tsv" % project
 
 
 wanted = """ERS091755
@@ -50,12 +57,13 @@ if not os.path.isfile(fastq_file):
 def process_fastq(project, fastq_filename):
     h = open(fastq_filename)
     line = h.readline()
-    assert line == 'study_accession\tsample_accession\tsecondary_sample_accession\texperiment_accession\trun_accession\ttax_id\tscientific_name\tinstrument_model\tlibrary_layout\tfastq_ftp\tfastq_galaxy\tsubmitted_ftp\tsubmitted_galaxy\tsra_ftp\tsra_galaxy\tcram_index_ftp\tcram_index_galaxy\n', repr(line)
+    assert line == "\t".join(fields) + "\n", repr(line)
     for line in h:
         parts = line.rstrip("\n").split("\t")
         assert parts[0] == project
-        urls = parts[9].split(";")
-        for url in urls:
+        urls = parts[fields.index("fastq_ftp")].split(";")
+        md5s = parts[fields.index("fastq_md5")].split(";")
+        for url, md5 in zip(urls, md5s):
             if url.startswith("ftp://ftp.sra.ebi.ac.uk/vol1/fastq/ERR"):
                 pass
             elif url.startswith("ftp.sra.ebi.ac.uk/vol1/fastq/ERR"):
@@ -66,9 +74,6 @@ def process_fastq(project, fastq_filename):
             if wanted and acc not in wanted:
                 print("Not interested in %s from %s" % (filename, acc))
                 continue
-            if os.path.isfile(filename):
-                print("Already have %s" % filename)
-                continue
             if not filename.endswith(".fastq.gz"):
                 print("Skipping %s" % filename)
                 continue
@@ -77,12 +82,22 @@ def process_fastq(project, fastq_filename):
             if not os.path.isdir(d):
                 print("Making directory %s" % d)
                 os.makedirs(d)
-            #Download file...
-            print("Downloading %s" % filename)
-            rc = os.system("wget -nv -O %s %s" % (filename, url))
-            assert not rc, rc
+            if os.path.isfile(filename):
+                print("Already have %s" % filename)
+            else:
+                #Download file...
+                print("Downloading %s" % filename)
+                rc = os.system("wget -nv -O %s %s" % (filename, url))
+                assert not rc, rc
             #Now check the md5...
-            print(filename)
+            m = filename + ".md5"
+            if not os.path.isfile(m):
+                print("Creating %s with md5 %s" % (m, md5))
+                with open(m, "w") as handle:
+                    handle.write("%s  %s" % (md5, os.path.basename(filename)))
+            print("Confirming %s has checksum %s" % (filename, md5))
+            rc = os.system("cd %s && md5sum -c %s" % (d, os.path.split(m)[1]))
+            assert not rc, rc
     h.close()
 
 process_fastq(project, fastq_file)
